SafePathZC/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.10
pydantic[email]==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6
requests==2.31.0
PyJWT==2.8.0104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.10
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6
requests==2.31.0
PyJWT==2.8.0
email-validator==2.1.0
